refsw: sync up to 12.06.2018.
refs: 1c44e37017fd2ade064ea8911cbfc7e083e37f24

+++ magnum/basemodules/dsp/src/arm/bdsp_arm_int_priv.c
@@ -85,7 +85,7 @@ void BDSP_Arm_P_AckEventCallback_isr(BTEE_ClientEvent event, void *pDeviceHandle
     int uiTaskIndex = 0;
 	BDSP_Arm_P_AckInfo sAckInfo;
 	BERR_Code ret = BERR_SUCCESS;
-	size_t ui32RcvMsgLen = 0;
+	uint32_t ui32RcvMsgLen = 0;
 	BTEE_ConnectionHandle hConnection = NULL;
 
 	BDBG_ENTER(BDSP_Arm_P_AckEventCallback_isr);
